+++ plugins/destination/postgresql/cloud-config-ui/src/utils/generateConnectionUrl.ts
@@ -1,26 +1,26 @@
 import { FormValues } from './formSchema';
 
 export function generateConnectionUrl(values: FormValues, replaceWithEnvs: boolean): string {
-  let { username } = values.spec;
+  let { username } = values;
 
   if (replaceWithEnvs) {
-    username = replaceWithEnvs ? (values.spec.username ? '${username}' : '') : values.spec.username;
+    username = replaceWithEnvs ? (values.username ? '${username}' : '') : values.username;
   }
 
-  const password = values.spec.password ? '${password}' : '';
+  const password = values.password ? '${password}' : '';
 
-  let finalUrl = `dbtype='postgresql' user='${username}' password='${password}' host='${escapeSingleQuotesAndBackslashes(values.spec.host)}' dbname='${escapeSingleQuotesAndBackslashes(values.spec.database)}'`;
+  let finalUrl = `dbtype='postgresql' user='${username}' password='${password}' host='${escapeSingleQuotesAndBackslashes(values.host)}' dbname='${escapeSingleQuotesAndBackslashes(values.database)}'`;
 
-  if (values.spec.port) {
-    finalUrl += ` port='${values.spec.port}'`;
+  if (values.port) {
+    finalUrl += ` port='${values.port}'`;
   }
 
-  if (values.spec.ssl) {
-    finalUrl += ` sslmode='${values.spec.sslMode}'`;
+  if (values.ssl) {
+    finalUrl += ` sslmode='${values.sslMode}'`;
   }
 
-  if (values.spec.clientEncoding) {
-    finalUrl += ` client_encoding='${escapeSingleQuotesAndBackslashes(values.spec.clientEncoding)}'`;
+  if (values.schemaName) {
+    finalUrl += ` search_path='${values.schemaName}'`;
   }
 
   return finalUrl;